chore/meta_voting: reorganize the module
- Rename modules to match the main type defined in it.
- Change visibility of some items from `pub` to `pub(crate)` to prevend accidentally exposing them to the outside.
- Do not ignore potential errors (add log_or_panic to a couple of places).
